fix(cli): address second round of review comments
- Extract resolve_grpc_socket_path into commands/mod.rs, replacing
  three identical copies in machine.rs, sandbox.rs, and migrate.rs
- Set allow_replacements=true unconditionally after user confirms,
  so blocker-only migrations (no resource replacements) also proceed
- Set success=true on intermediate progress events to avoid misleading
  clients that check success without gating on done
