+++ app/arcbox-cli/src/commands/machine.rs
@@ -20,37 +20,8 @@ use tokio::net::UnixStream;
 use tonic::codegen::{Service, http::Uri};
 use tonic::transport::{Channel, Endpoint};
 
-fn resolve_grpc_socket_path() -> PathBuf {
-    if let Ok(path) = std::env::var("ARCBOX_GRPC_SOCKET") {
-        return PathBuf::from(path);
-    }
-
-    if let Ok(path) = std::env::var("ARCBOX_SOCKET") {
-        let docker_socket = PathBuf::from(path);
-        if let Some(parent) = docker_socket.parent() {
-            let preferred = parent.join("arcbox-grpc.sock");
-            if preferred.exists() {
-                return preferred;
-            }
-
-            let legacy = parent.join("arcbox.sock");
-            if legacy.exists() {
-                return legacy;
-            }
-
-            return preferred;
-        }
-    }
-
-    dirs::home_dir()
-        .unwrap_or_else(|| PathBuf::from("/tmp"))
-        .join(".arcbox")
-        .join(arcbox_constants::paths::host::RUN)
-        .join("arcbox.sock")
-}
-
 async fn machine_client() -> Result<MachineServiceClient<Channel>> {
-    let socket_path = resolve_grpc_socket_path();
+    let socket_path = super::resolve_grpc_socket_path();
 
     let channel = Endpoint::from_static("http://[::]:50051")
         .connect_with_connector(UnixConnector::new(socket_path.clone()))
