+++ app/arcbox-cli/src/commands/mod.rs
@@ -14,6 +14,38 @@
 use clap::{Parser, Subcommand, ValueEnum};
 use std::path::PathBuf;
 
+/// Resolves the gRPC socket path from environment or default location.
+///
+/// Shared by machine, sandbox, and migrate command modules.
+pub fn resolve_grpc_socket_path() -> PathBuf {
+    if let Ok(path) = std::env::var("ARCBOX_GRPC_SOCKET") {
+        return PathBuf::from(path);
+    }
+
+    if let Ok(path) = std::env::var("ARCBOX_SOCKET") {
+        let docker_socket = PathBuf::from(path);
+        if let Some(parent) = docker_socket.parent() {
+            let preferred = parent.join("arcbox-grpc.sock");
+            if preferred.exists() {
+                return preferred;
+            }
+
+            let legacy = parent.join("arcbox.sock");
+            if legacy.exists() {
+                return legacy;
+            }
+
+            return preferred;
+        }
+    }
+
+    dirs::home_dir()
+        .unwrap_or_else(|| PathBuf::from("/tmp"))
+        .join(".arcbox")
+        .join(arcbox_constants::paths::host::RUN)
+        .join("arcbox.sock")
+}
+
 pub mod boot;
 pub mod daemon;
 #[cfg(target_os = "macos")]
